litellm_chainlit_uv/pyproject.toml
[project]
name = "litellm-chainlit-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "google-auth>=2.40.2",
    "litellm>=1.71.1",
    "python-dotenv>=1.1.0",
]
